# attendance-backend-main/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CheckInRequest": {
				"properties": {
					"visitor": {
						"properties": {
							"idNumber": {
								"type": "string"
							},
							"idProofType": {
								"type": "string"
							},
							"hostName": {
								"type": "string"
							},
							"department": {
								"type": "string"
							},
							"purpose": {
								"type": "string"
							},
							"visitorCompany": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"mobile": {
								"type": "string"
							},
							"fullName": {
								"type": "string"
							}
						},
						"required": [
							"fullName"
						],
						"type": "object"
					},
					"verificationMode": {
						"type": "string"
					},
					"appointmentId": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"badgeType": {
						"type": "string",
						"enum": [
							"QR",
							"RFID"
						]
					}
				},
				"required": [
					"visitor"
				],
				"type": "object",
				"additionalProperties": {}
			},
...
			"UserStatus": {
				"type": "string",
				"enum": [
					"active",
					"inactive",
					"suspended",
					"pending",
					"deleted",
					"archived",
					"rejected",
					"blacklisted"
				]
			},
			"IPermissionAttributes": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IRoleAttributes": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"permissions": {
						"items": {
							"$ref": "#/components/schemas/IPermissionAttributes"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IUserAttributes": {
				"properties": {
					"id": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phoneNumber": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"status": {
						"$ref": "#/components/schemas/UserStatus"
					},
					"roles": {
						"items": {
							"$ref": "#/components/schemas/IRoleAttributes"
						},
						"type": "array"
					},
					"profilePicture": {
						"type": "string"
					},
					"dateOfBirth": {
						"type": "string",
						"format": "date-time"
					},
					"nationalId": {
						"type": "string"
					},
					"fingerPrint": {
						"type": "string"
					},
					"face": {
						"type": "string"
					},
					"voice": {
						"type": "string"
					},
					"signature": {
						"type": "string"
					},
					"passport": {
						"type": "string"
					},
					"motion": {
						"type": "string"
					},
					"igipande": {
						"type": "string"
					},
					"ocr": {
						"type": "string"
					},
					"gesture": {
						"type": "string"
					},
					"pupil": {
						"type": "string"
					},
					"otherBiometric": {
						"type": "string"
					},
					"nationality": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"sector": {
						"type": "string"
					},
					"cell": {
						"type": "string"
					},
					"village": {
						"type": "string"
					},
					"personalInChargeNationalId": {
						"type": "string"
					},
					"personalInChargeName": {
						"type": "string"
					},
					"personalInChargePhone": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"personalInChargeEmail": {
						"type": "string"
					},
					"personalInChargeRelation": {
						"type": "string"
					},
					"personalInChargeOtherDetails": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"fullName",
					"email",
					"password",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UsersListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/IUserAttributes"
						},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IUserAttributes"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AttendanceInfo": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"checkIn": {
						"type": "string",
						"format": "date-time"
					},
					"date": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"userId",
					"checkIn",
					"date",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserWithAttendanceResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"properties": {
							"assignedCardId": {
								"type": "string",
								"nullable": true
							},
							"equipmentIds": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"isNewUser": {
								"type": "boolean"
							},
							"attendance": {
								"allOf": [
									{
										"$ref": "#/components/schemas/AttendanceInfo"
									}
								],
								"nullable": true
							},
							"user": {
								"allOf": [
									{
										"$ref": "#/components/schemas/IUserAttributes"
									}
								],
								"nullable": true
							}
						},
						"required": [
							"assignedCardId",
							"equipmentIds",
							"isNewUser",
							"attendance",
							"user"
						],
						"type": "object",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
			"EquipmentAssignment": {
				"properties": {
					"name": {
						"type": "string"
					},
					"serialNumber": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"serialNumber"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserCreateRequest": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"scannedId": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"nationalId": {
						"type": "string"
					},
					"equipments": {
						"items": {
							"$ref": "#/components/schemas/EquipmentAssignment"
						},
						"type": "array"
					},
					"cardId": {
						"type": "string"
					}
				},
				"required": [
					"fullName",
					"scannedId"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UserUpdateRequest": {
				"properties": {
					"email": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"company": {
						"type": "string"
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
...
			"AttendanceResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"checkIn": {
						"type": "string",
						"format": "date-time"
					},
					"checkOut": {
						"type": "string",
						"format": "date-time"
					},
					"hoster": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"user": {},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"checkIn",
					"date",
					"status"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AttendanceListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/AttendanceResponse"
						},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"AppointmentResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"purpose": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"cancelReason": {
						"type": "string"
					},
					"reasonToReschedule": {
						"type": "string"
					},
					"appointmentDate": {
						"type": "string"
					},
					"appointmentTime": {
						"type": "string"
					},
					"user": {
						"properties": {
							"phoneNumber": {
								"type": "string"
							},
							"department": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"fullName": {
								"type": "string"
							}
						},
						"required": [
							"phoneNumber",
							"department",
							"fullName"
						],
						"type": "object"
					},
					"timeDuration": {
						"type": "string"
					},
					"appointmentLocation": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"user"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ServiceResponse_AppointmentResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/AppointmentResponse"
						},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT token (Bearer <token>)"
			}
		}
	},
	"info": {
		"title": "Attendance Backend API",
		"version": "1.0.0",
		"description": "API documentation for attendance management system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/visitors/check-in": {
			"post": {
				"operationId": "CheckIn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckInRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UsersListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:list"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserWithAttendanceResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/api/users/{id}/equipment": {
			"get": {
				"operationId": "GetUserEquipment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/{id}/activate": {
			"put": {
				"operationId": "ActivateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/{id}/suspend": {
			"put": {
				"operationId": "SuspendUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/users/{id}/delete": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/attendance/report/user/{userId}/{startDate}/{endDate}": {
			"get": {
				"operationId": "GetUserAttendanceReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/appointments/user/{userId}": {
			"get": {
				"operationId": "GetAppointmentsByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_AppointmentResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Appointments"
				],
				"security": [
					{
						"jwt": [
							"appointment:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}